Add support for service accounts API
Added new endpoints for service accounts and service account keys.
Also fixed an issue with optional parameters in object style.

ex:

```
smooch.apps.list({
    serviceAccountId: 'my_id'
})
```

would throw an error `limit must be a number`, because it was passing
`serviceAccountId` as the first arg to the `list` call, instead of
passing `undefined` for the `limit` and `offset` params, and then
passing `serviceAccountId` as 3rd param.